# apps/go-backend/application/dto/book_dto.go
package dto

import "time"

// BookDTO is the API representation of a Book
type BookDTO struct {
	ID        uint      `json:"id"`
	Title     string    `json:"title"`
	Author    string    `json:"author"`
	Year      int       `json:"year"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

// BooksResponse is the paginated response for books list
type BooksResponse struct {
	Data       []BookDTO `json:"data"`
	Total      int64     `json:"total"`
	Page       int       `json:"page"`
	PageSize   int       `json:"pageSize"`
	TotalPages int       `json:"totalPages"`
}